Refactor iframe communication and use MessageChannel
